requirements.txt
aws_cdk.core

aws_cdk.aws-ec2